api/api_marker/pyproject.toml
[project]
name = "api-marker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "extraction-common",
    "fastapi>=0.136.0",
    "langfuse>=4.5.1",
    "marker-pdf>=1.10.2",
    "pillow>=10.4.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.45.0",
]

[tool.uv.sources]
extraction-common = { path = "../../libs" }
